# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "StablePay API",
        "description": "Cross-border stablecoin remittance API on Solana. Enables instant, low-cost USD to INR remittances via USDC with MPC wallet abstraction and guaranteed delivery.",
        "contact": {
            "name": "StablePay Team",
            "url": "https://github.com/stablepay"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/remittances": {
            "get": {
                "tags": [
                    "Remittances"
                ],
                "summary": "List remittances",
                "description": "Returns a paginated list of remittances for the authenticated user",
                "operationId": "listRemittances",
                "parameters": [
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Remittances retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageRemittanceResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Remittances"
                ],
                "summary": "Create remittance",
                "description": "Initiates a new USD to INR remittance with locked FX rate",
                "operationId": "createRemittance",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRemittanceRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or insufficient balance",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Remittance created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RemittanceResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/remittances/{remittanceId}": {
            "get": {
                "tags": [
                    "Remittances"
                ],
                "summary": "Get remittance",
                "description": "Retrieves a remittance by its unique identifier",
                "operationId": "getRemittance",
                "parameters": [
                    {
                        "name": "remittanceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Remittance not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Remittance found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RemittanceResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/remittances/{remittanceId}/timeline": {
            "get": {
                "tags": [
                    "Remittances"
                ],
                "summary": "Get remittance timeline",
                "description": "Returns a 4-step progress timeline for the remittance",
                "operationId": "getTimeline",
                "parameters": [
                    {
                        "name": "remittanceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Remittance not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Timeline retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RemittanceTimelineResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "errorCode": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "path": {
                        "type": "string"
                    }
                }
            },
...
            "CreateRemittanceRequest": {
                "type": "object",
                "properties": {
                    "recipientPhone": {
                        "type": "string",
                        "minLength": 1
                    },
                    "amountUsdc": {
                        "type": "number"
                    }
                },
                "required": [
                    "amountUsdc",
                    "recipientPhone"
                ]
            },
            "RemittanceResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "remittanceId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "recipientPhone": {
                        "type": "string"
                    },
                    "amountUsdc": {
                        "type": "number"
                    },
                    "amountInr": {
                        "type": "number"
                    },
                    "fxRate": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "INITIATED",
                            "ESCROWED",
                            "CLAIMED",
                            "DELIVERED",
                            "DISBURSEMENT_FAILED",
                            "DEPOSIT_FAILED",
                            "CLAIM_FAILED",
                            "REFUND_FAILED",
                            "REFUNDED",
                            "CANCELLED"
                        ]
                    },
                    "escrowPda": {
                        "type": "string"
                    },
                    "claimTokenId": {
                        "type": "string"
                    },
                    "smsNotificationFailed": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "PageRemittanceResponse": {
                "type": "object",
                "properties": {
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RemittanceResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "PageableObject": {
                "type": "object",
                "properties": {
                    "offset": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "unpaged": {
                        "type": "boolean"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "paged": {
                        "type": "boolean"
                    },
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "SortObject": {
                "type": "object",
                "properties": {
                    "empty": {
                        "type": "boolean"
                    },
                    "unsorted": {
                        "type": "boolean"
                    },
                    "sorted": {
                        "type": "boolean"
                    }
                }
            },
            "RemittanceTimelineResponse": {
                "type": "object",
                "properties": {
                    "steps": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TimelineStep"
                        }
                    },
                    "failed": {
                        "type": "boolean"
                    }
                }
            },
            "TimelineStep": {
                "type": "object",
                "properties": {
                    "step": {
                        "type": "string",
                        "enum": [
                            "INITIATED",
                            "ESCROWED",
                            "CLAIMED",
                            "DELIVERED",
                            "DISBURSEMENT_FAILED",
                            "DEPOSIT_FAILED",
                            "CLAIM_FAILED",
                            "REFUND_FAILED",
                            "REFUNDED",
                            "CANCELLED"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "COMPLETED",
                            "CURRENT",
                            "PENDING",
                            "FAILED"
                        ]
                    },
                    "message": {
                        "type": "string"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}